+++ app/src/main/java/org/sopt/at/feature/main/MainViewModel.kt
@@ -1,0 +1,10 @@
+package org.sopt.at.feature.main
+
+import androidx.lifecycle.ViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+
+class MainViewModel : ViewModel() {
+
+}
