+++ app/src/main/java/org/sopt/at/feature/main/search/SearchScreen.kt
@@ -1,0 +1,25 @@
+package org.sopt.at.feature.main.search
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun SearchScreen(modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text("검색 화면")
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun PreviewSearchScreen() {
+    SearchScreen()
+}
