--- app/src/main/java/org/sopt/at/presentation/ui/main/MainViewModel.kt
@@ -1,10 +1,0 @@
-package org.sopt.at.presentation.ui.main
-
-import androidx.lifecycle.ViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-
-class MainViewModel : ViewModel() {
-
-}
