--- app/src/main/java/org/sopt/at/presentation/ui/signup/content/SignUpContent.kt
@@ -1,191 +1,0 @@
-package org.sopt.at.presentation.ui.signup.content
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.imePadding
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import org.sopt.at.R
-import org.sopt.at.core.component.button.TvingButton
-import org.sopt.at.core.component.textfield.IdTextField
-import org.sopt.at.core.component.textfield.PasswordTextField
-import org.sopt.at.core.component.topbar.TvingTopBar
-import org.sopt.at.ui.theme.TvingTheme
-
-@Composable
-fun SignUpContent(
-    modifier: Modifier = Modifier,
-    step: Int,
-    id: String = "",
-    onIdChange: (String) -> Unit = {},
-    pw: String = "",
-    onPwChange: (String) -> Unit = {},
-    nickname: String = "",
-    onNicknameChange: (String) -> Unit = {},
-    onNext: () -> Unit,
-    onBack: () -> Unit,
-    showErrorSnackbar: (String) -> Unit
-) {
-    var showPassword by remember { mutableStateOf(value = false) }
-
-    val isValid = when (step) {
-        1 -> nickname.matches(Regex("^[가-힣a-zA-Z0-9]{1,20}$"))
-        2 -> id.matches(Regex("^[A-Za-z0-9]{8,20}$"))
-        else -> pw.matches(Regex("^[A-Za-z0-9]{8,20}$"))
-    }
-
-    val invalidNicknameMsg = stringResource(R.string.sign_up_invalid_nickname)
-    val invalidIdMsg = stringResource(R.string.sign_up_invalid_id)
-    val invalidPwMsg = stringResource(R.string.sign_up_invalid_pw)
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(color = TvingTheme.colors.BasicBlack),
-    ) {
-        TvingTopBar(
-            onBackClick = onBack
-        )
-
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(15.dp)
-                .imePadding(),
-            verticalArrangement = Arrangement.SpaceBetween
-        ) {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                // 제목
-                Text(
-                    text = when (step) {
-                        1 -> stringResource(R.string.sign_up_nickname_input)
-                        2 -> stringResource(R.string.sign_up_id_input)
-                        else -> stringResource(R.string.sign_up_pw_input)
-                    },
-                    color = TvingTheme.colors.BasicWhite,
-                    style = TvingTheme.typography.subtitle
-                )
-
-                when (step) {
-                    1 -> {
-                        // 닉네임 입력 창
-                        Column {
-                            IdTextField(
-                                id = nickname,
-                                onIdChange = onNicknameChange,
-                                placeholderText = stringResource(R.string.textfield_nickname),
-                                modifier = Modifier
-                                    .padding(top = 20.dp)
-                                    .border(
-                                        width = 1.dp,
-                                        color = TvingTheme.colors.Gray2,
-                                        shape = RoundedCornerShape(5.dp)
-                                    )
-                            )
-
-                            Text(
-                                text = stringResource(R.string.sign_up_nickname_condition),
-                                color = TvingTheme.colors.Gray2,
-                                style = TvingTheme.typography.caption
-                            )
-                        }
-                    }
-
-                    2 -> {
-                        // 아이디 입력 창
-                        Column {
-                            IdTextField(
-                                id = id,
-                                onIdChange = onIdChange,
-                                placeholderText = stringResource(R.string.textfield_id),
-                                modifier = Modifier
-                                    .padding(top = 20.dp)
-                                    .border(
-                                        width = 1.dp,
-                                        color = TvingTheme.colors.Gray2,
-                                        shape = RoundedCornerShape(5.dp)
-                                    )
-                            )
-
-                            Text(
-                                text = stringResource(R.string.sign_up_id_condition),
-                                color = TvingTheme.colors.Gray2,
-                                style = TvingTheme.typography.caption
-                            )
-                        }
-                    }
-
-                    else -> {
-                        // 비밀번호 입력 창
-                        Column {
-                            PasswordTextField(
-                                password = pw,
-                                onPasswordChange = onPwChange,
-                                showPassword = showPassword,
-                                onTogglePasswordVisibility = { showPassword = !showPassword },
-                                modifier = Modifier
-                                    .padding(top = 20.dp)
-                                    .border(1.dp, Color.Gray, shape = RoundedCornerShape(5.dp)),
-                            )
-
-                            Text(
-                                text = stringResource(R.string.sign_up_pw_condition),
-                                color = TvingTheme.colors.Gray2,
-                                style = TvingTheme.typography.caption
-                            )
-                        }
-                    }
-                }
-            }
-
-            // 다음 버튼
-            TvingButton(
-                onClick = {
-                    if (isValid) {
-                        onNext()
-                    } else {
-                        val message = when (step) {
-                            1 -> invalidNicknameMsg
-                            2 -> invalidIdMsg
-                            else -> invalidPwMsg
-                        }
-                        showErrorSnackbar(message)
-                    }
-                },
-                content = stringResource(R.string.button_next)
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun PreviewSignUpContent() {
-    SignUpContent(
-        step = 1,
-        id = "",
-        onIdChange = {},
-        pw = "",
-        onPwChange = {},
-        onNext = {},
-        onBack = {},
-        showErrorSnackbar = {}
-    )
-}
