--- app/src/main/java/org/sopt/at/presentation/ui/splash/SplashViewModel.kt
@@ -1,32 +1,0 @@
-package org.sopt.at.presentation.ui.splash
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-import org.sopt.at.data.local.AuthPreferences
-import javax.inject.Inject
-
-@HiltViewModel
-class SplashViewModel @Inject constructor(
-    private val authPreferences: AuthPreferences
-) : ViewModel() {
-
-    private val _isLoggedIn = MutableStateFlow<Boolean?>(null)
-    val isLoggedIn: StateFlow<Boolean?> = _isLoggedIn
-
-    init {
-        checkLoginState()
-    }
-
-    private fun checkLoginState() {
-        viewModelScope.launch {
-            authPreferences.userId
-                .collect { id ->
-                    _isLoggedIn.value = id != 0L
-                }
-        }
-    }
-}
